+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@
 ## New Features
 
 * Add helper function to support creating quantities that conform to the API expectations.
+* Add `receive-gridpool-trades` command to CLI tool.
 
 <!-- Here goes the main new features and examples or instructions on how to use them -->